+++ src/polyfills.js
@@ -10,6 +10,12 @@
 	};
 })();
 
+if (!String.prototype.trim) {
+	String.prototype.trim = function () {
+		return this.replace(/^\s+|\s+$/g, '');
+	};
+}
+
 if(!Array.isArray) {
 	/**
 	 * Array.isArray dle ES5 - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/isArray